lasers ig swing or be swung

+++ BULLET HELL/Assets/Scripts/Enemy/MiddleWare/Enemy_Shootpoint1Command.cs
@@ -91,7 +91,7 @@ public class Enemy_Shoot1Command : MonoBehaviour
     {
         if (nullNeeded)
             actionNull();
-        shootingPattern.setFireRate((int)(desired_FireRate / 2));
+        shootingPattern.setFireRate((desired_FireRate / 2));
         shootingPattern.setCanShoot(true);
     }
