+++ BULLET HELL/Assets/Scripts/Enemy/MiddleWare/Enemy_Shootpoint2Command.cs
@@ -90,7 +90,7 @@ public class Enemy_Shoot2Command : MonoBehaviour
     {
         if (nullNeeded)
             actionNull();
-        shootingPattern.setFireRate((int)(desired_FireRate / 2));
+        shootingPattern.setFireRate((desired_FireRate / 2));
         shootingPattern.setCanShoot(true);
     }
