+++ BULLET HELL/Assets/Scripts/Enemy/MiddleWare/Enemy_ShootpointAimCommand.cs
@@ -95,7 +95,7 @@ public class Enemy_ShootAimCommand : MonoBehaviour
     {
         if (nullNeeded)
             actionNull();
-        shootingPattern.setFireRate((int)(desired_FireRate / 2));
+        shootingPattern.setFireRate((desired_FireRate / 2));
         shootingPattern.setCanShoot(true);
     }
